+++ R/hits.R
@@ -20,6 +20,8 @@ NULL
 #'   `check_cqp_query`.
 #' @param s_attribute A `character` vector of s-attributes that will be
 #'   used to breakdown counts for matches for query/queries.
+#' @param decode A `logical` value, whether to decode s-attributes. If `FALSE`,
+#'   the `integer` values of strucs are reported in the table with matches.
 #' @param p_attribute A `character` vector stating a p-attribute.
 #' @param size A `logical` value, whether to report the size of subcorpus.
 #' @param freq A `logcial` value, whether to report relative frequencies.
@@ -59,9 +61,19 @@ setGeneric("hits", function(.Object, ...) standardGeneric("hits"))
 #' y <- corpus("REUTERS") %>%
 #'   subset(grep("saudi-arabia", places)) %>%
 #'   hits(query = "oil")
-setMethod("hits", "corpus", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, fill = FALSE, mc = 1L, verbose = TRUE, progress = FALSE, ...){
+setMethod("hits", "corpus", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, decode = TRUE, fill = FALSE, mc = 1L, verbose = TRUE, progress = FALSE, ...){
   
   if (is.logical(mc)) if (mc) mc <- getOption("polmineR.cores") else mc <- 1L
+  
+  stopifnot(is.logical(decode))
+  if (!missing(s_attribute)){
+    if (length(decode) == 1L){
+      decode <- rep(decode, times = length(s_attribute))
+    } else {
+      if (length(decode) != length(s_attribute))
+        stop("Length of vectors or argument `decode` and `s_attribute` differ")
+    }
+  }
   
   if ("sAttribute" %in% names(list(...))){
     lifecycle::deprecate_warn(
@@ -115,11 +127,17 @@ setMethod("hits", "corpus", function(.Object, query, cqp = FALSE, check = TRUE, 
       corpus = .Object@corpus, registry = .Object@registry_dir,
       s_attribute = s_attribute[i], cpos = DT[["cpos_left"]]
     )
-    s_attr_values <- cl_struc2str(
-      corpus = .Object@corpus, registry = .Object@registry_dir,
-      s_attribute = s_attribute[i], struc = strucs
-    )
-    DT[, eval(s_attribute[i]) := as.nativeEnc(s_attr_values, from = .Object@encoding)]
+    if (!s_attr_has_values(s_attribute[i], x = .Object)) decode[i] <- FALSE
+    if (decode[i]){
+      s_attr_values <- cl_struc2str(
+        corpus = .Object@corpus, registry = .Object@registry_dir,
+        s_attribute = s_attribute[i], struc = strucs
+      ) |>
+        as.nativeEnc(from = .Object@encoding)
+      DT[, eval(s_attribute[i]) := s_attr_values]
+    } else {
+      DT[, eval(s_attribute[i]) := strucs]
+    }
   }
   
   TF <- DT[, .N, by = c(eval(c("query", s_attribute))), with = TRUE]
@@ -162,26 +180,33 @@ setMethod("hits", "corpus", function(.Object, query, cqp = FALSE, check = TRUE, 
 
 
 #' @rdname hits
-setMethod("hits", "character", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, mc = FALSE, verbose = TRUE, progress = FALSE, ...){
+setMethod("hits", "character", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, decode = TRUE, mc = FALSE, verbose = TRUE, progress = FALSE, ...){
   hits(
     .Object = corpus(.Object),
-    query = query, cqp = cqp, check = check,
-    s_attribute = s_attribute, p_attribute = p_attribute,
-    size = size, freq = freq,
-    mc = mc, verbose = verbose, progress = progress,
+    query = query,
+    cqp = cqp,
+    check = check,
+    s_attribute = s_attribute,
+    p_attribute = p_attribute,
+    size = size,
+    freq = freq,
+    decode = decode,
+    mc = mc,
+    verbose = verbose,
+    progress = progress,
     ...)
 })
 
 
 
 #' @rdname hits
-setMethod("hits", "subcorpus", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, fill = FALSE, mc = FALSE, progress = FALSE, verbose = TRUE, ...){
+setMethod("hits", "subcorpus", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, fill = FALSE, decode = TRUE, mc = FALSE, progress = FALSE, verbose = TRUE, ...){
   callNextMethod()
 })
 
 
 #' @rdname hits
-setMethod("hits", "partition", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, fill = FALSE, mc = FALSE, progress = FALSE, verbose = TRUE, ...){
+setMethod("hits", "partition", function(.Object, query, cqp = FALSE, check = TRUE, s_attribute, p_attribute = "word", size = FALSE, freq = FALSE, fill = FALSE, decode = TRUE, mc = FALSE, progress = FALSE, verbose = TRUE){
   hits(
     .Object = as(.Object, "subcorpus"),
     query = query,
@@ -192,6 +217,7 @@ setMethod("hits", "partition", function(.Object, query, cqp = FALSE, check = TRU
     size = size, 
     freq = freq,
     fill = fill,
+    decode = decode,
     mc = mc, 
     progress = progress,
     verbose